Add CheckpointTimer to simplify time measurement in loops
The `CheckpointTimer` class is intended for time measurement of loops
where the time spent in separate sections of the code is measured, as
well as the total duration of the loop.  See doxygen of the class for
more details.

It is basically a wrapper around a list of `Timer` instances with the
purpose to not clutter the measured code too much.  Also it can easily
be disabled via a template argument which should result in the compiler
optimizing away the function calls completely (i.e. no need to have
commented lines in the code).